requirements.txt
anyio==4.12.1
certifi==2026.1.4
exceptiongroup==1.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
python-dotenv==1.2.1
python-telegram-bot==21.6
typing_extensions==4.15.0
